srv: move camera handler to separate module, implement picamera with basic functionality
